Move icons into icon directory
