+++ src/icons/Auto.tsx
