+++ src/icons/Light.tsx
